+++ app/components/listings/ListingInfo.tsx
@@ -15,12 +15,14 @@ interface ListingInfoProps {
         description: string;
       }
     | undefined;
+  ownerIsWatching: boolean;
 }
 
 const ListingInfo: React.FC<ListingInfoProps> = ({
   user,
   description,
   category,
+  ownerIsWatching,
 }) => {
   const { t } = useTranslation();
 
@@ -37,9 +39,15 @@ const ListingInfo: React.FC<ListingInfoProps> = ({
             gap-2
           "
         >
-          <div>
-            {t("The_listing_was_posted_by")} {user?.firstName} {user?.lastName}
-          </div>
+          {ownerIsWatching ? (
+            <div>{t("You_are_now_looking_at_your_own_listing")}</div>
+          ) : (
+            <div>
+              {t("The_listing_was_posted_by")} {user?.firstName}{" "}
+              {user?.lastName}
+            </div>
+          )}
+
           <Avatar width={60} height={60} src={user?.image} />
         </div>
         <div